lat/lon, float to decimal

+++ migrations/20241110193147_allow.js
@@ -4,11 +4,11 @@
  */
 exports.up = function (knex) {
     return knex.schema.alterTable('open_cities', (table) => {
-        table.float('bbox_lat_min', 8, 4).nullable().alter();
-        table.float('bbox_lat_max', 8, 4).nullable().alter();
+        table.decimal('bbox_lat_min', 8, 4).nullable().alter();
+        table.decimal('bbox_lat_max', 8, 4).nullable().alter();
 
-        table.float('bbox_lon_min', 8, 4).nullable().alter();
-        table.float('bbox_lon_max', 8, 4).nullable().alter();
+        table.decimal('bbox_lon_min', 8, 4).nullable().alter();
+        table.decimal('bbox_lon_max', 8, 4).nullable().alter();
 
         table.mediumint('bbox_lat_min_1000').nullable().alter();
         table.mediumint('bbox_lat_max_1000').nullable().alter();
